+++ R/CG_inflator.R
@@ -25,7 +25,7 @@ CG_population_inflator <- function(x = 1,
   nse_forecast_series <- forecast.series
   
   out_tbl <- 
-    cg_inflators_1314 %>% 
+    cg_inflators_1516 %>% 
     copy %>%
     .[forecast.series == nse_forecast_series]
   
@@ -45,9 +45,14 @@ CG_inflator <- function(x = 1, from_fy, to_fy, forecast.series = "mean"){
   prohibit_vector_recycling(x, from_fy, to_fy)
   stopifnot(is.numeric(x), all(is.fy(from_fy)), all(is.fy(to_fy)))
   
+  nse_forecast_series <- forecast.series
+  cg_inflators_tbl <- 
+    cg_inflators_1516[forecast.series == nse_forecast_series]
+  
+  
   # Else NAs.
-  stopifnot(all(to_fy %in% cg_inflators_1314[["fy_year"]]),
-            all(from_fy %in% cg_inflators_1314[["fy_year"]]))
+  stopifnot(all(to_fy %in% cg_inflators_1516[["fy_year"]]),
+            all(from_fy %in% cg_inflators_1516[["fy_year"]]))
 
   # CRAN Note avoidance
   ordering <- NULL
@@ -55,11 +60,7 @@ CG_inflator <- function(x = 1, from_fy, to_fy, forecast.series = "mean"){
     data.table(x = x, from_fy = from_fy, to_fy = to_fy) %>% 
     .[, ordering := 1:.N]
 
-  nse_forecast_series <- forecast.series
-  cg_inflators_tbl <- 
-    cg_inflators_1314 %>% 
-    copy %>% 
-    .[forecast.series == nse_forecast_series]
+  
   
   raw_out <- 
     input %>%